// public/tradingview/charting_library/charting_library.d.ts
export declare type Nominal<T, Name extends string> = T & {
  /* eslint-disable-next-line jsdoc/require-jsdoc */ [Symbol.species]: Name
}
// ...
export interface IExternalDatafeed {
  /**
   * This call is intended to provide the object filled with the configuration data.
   * The lib assumes that you will call the callback function and pass your datafeed {@link DatafeedConfiguration} as an argument.
   *
   * @param  {OnReadyCallback} callback - callback to return your datafeed configuration ({@link DatafeedConfiguration}) to the library.
   */
  onReady(callback: OnReadyCallback): void
}
// ...
export interface Mark {
  /** ID of the mark */
  id: string | number
  /**
   * Time for the mark.
   * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
   */
  time: number
  /** Color for the mark */
  color: MarkConstColors | MarkCustomColor
  /** Text content for the mark */
  text: string
  /** Label for the mark */
  label: string
  /** Text color for the mark */
  labelFontColor: string
  /** Minimum size for the mark */
  minSize: number
  /** Border Width */
  borderWidth?: number
  /** Border Width when hovering over bar mark */
  hoveredBorderWidth?: number
  /**
   * Optional URL for an image to be displayed within the timescale mark.
   *
   * The image should ideally be square in dimension. You can use any image type which
   * the browser supports natively.
   *
   * Examples:
   * - `https://yourserver.com/adobe.svg`
   * - `/images/myImage.png`
   * - `data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3...`
   * - `data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAASABIAAD/4...`
   */
  imageUrl?: string
  /**
   * Continue to show text label even when an image has
   * been loaded for the timescale mark.
   *
   * Defaults to `false` if undefined.
   */
  showLabelWhenImageLoaded?: boolean
}
export interface MarkCustomColor {
  /** Border color */
  border: string
  /** Background color */
  background: string
}
// ...
export type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed
// ...
export type MarkConstColors = 'red' | 'green' | 'blue' | 'yellow'
// ...
export type OnReadyCallback = (configuration: DatafeedConfiguration) => void
// ...
export type ResolutionString = Nominal<string, 'ResolutionString'>